\name{chernoff_realizations}
\alias{chernoff_realizations}
\docType{data}
\title{Quantiles of the Chernoff Random Variable}
\description{Quantiles of the Chernoff Random Variable that 
are used within the wald-type confidence interval functions.
}
\usage{data(chernoff_realizations)}
\format{A data frame of length 200 with quantiles and density values.
}
\references{
Computing Chernoff's Distribution
Piet Groeneboom, Jon A Wellner 
Journal of Computational and Graphical Statistics 
Vol. 10, Iss. 2, 2001
}
\examples{
data(chernoff_realizations)
}
\keyword{datasets}

