\name{RVforLR_realizations}
\alias{RVforLR_realizations}
\docType{data}
\title{Realizations of Random variable for LR-based confidence intervals}
\description{Realizations of the random variable that the 
likelihood ratio test statistic converges to}
\usage{data("RVforLR_realizations")}
\format{A data frame of realizations.}
\references{
Banerjee M (2000). Likelihood Ratio Inference in Regular and Non-regular Problems. Ph.D.
thesis, University of Washington.

Banerjee M (2009). 'Inference in exponential family regression models under certain shape
constraints.' In Advances in Multivariate Statistical Methods, Statistical Science and 
Interdisciplinary Research, volume 4, pp. 249-72. World Scientific.

Banerjee M, Wellner J (2001). 'Likelihood Ratio Tests for Monotone Functions.' Annals of
Statistics, 29, 1699 - 1731.
}
\examples{
data("RVforLR_realizations")
}
\keyword{datasets}

